+++ NEWS.md
@@ -9,7 +9,9 @@
 * New `isSorted()` method to properly test with `ignore.strand` set to `TRUE`.
 * Default `plotApairOfChrs()` to the first _sequence level_ of the `GBreaks` object.
 * `plotApairOfChrs()` now renders regions as blocks and not arrows as the strand
-  information is encoded in the color and shape of the comparison bands.
+  information is encoded in the color and shape of the comparison bands. It
+  also plots in light blue the blocks that are not matched.  Closes issues #3
+  and #5.
 * Added example data, see ?StructuralVariants for details.
 
 ## Bug fixes